Updated styles significantly
