--- src/map-interface/map-page/transitions.module.styl
@@ -1,50 +1,0 @@
-core_transition=opacity 0.8s ease, height 0.8s ease, max-height 0.8s ease
-
-opacity-transition-core(child_class, prefix)
-  {child_class}
-    transition: core_transition
-  &{prefix}from {child_class}
-    opacity: 0
-  &{prefix}enter {child_class}
-    opacity: 1
-  &{prefix}leave {child_class}
-    opacity: 0
-
-opacity-transition(child_class)
-  opacity-transition-core(child_class, child_class + '-')
-
-height-transition(child_class, prefix)
-  {child_class}
-    transition: core_transition
-  &{prefix}from {child_class}
-    max-height: 0
-    height: 0
-  &{prefix}leave {child_class}
-    max-height: 0
-    max-height: 0
-
-.main-ui
-  opacity-transition('.detail-panel')
-  opacity-transition('.context-panel')
-
-@media only screen and (max-width: 768px)
-  .main-ui
-    height-transition('.detail-stack', '.detail-panel-')
-    height-transition('.context-panel', '.context-panel-')
-
-@media only screen and (min-width: 768px)
-
-  .map-page .main-ui.detail-panel-leave .map-view-container
-    margin-right -14em
-
-  .map-page
-    &.map-is-global
-      .main-ui
-        &.detail-panel-leave .map-view-container
-          margin-right -30em
-          //margin-left: -30em
-
-        // Adjust so that the map centers on a further left
-        // region of the screen when the context panel is not open.
-        &.context-panel-leave .map-view-container
-          margin-left: -16em
